Crate prjunnamed_memory

This library contains various common utilities for writing target-specific memory lowering passes.

Target-specific memory lowering passes are organized roughly as follows:

for each memory in the design:

a list of candidate lowerings is assembled:

if a lowering mode (ff / lutram / blockram / hugeram) is explicitly selected, only the relevant target cells types are considered

for each available type of target cell:

hard requirements are checked, and the target cell is rejected if they are unmet, eg:

memories with more write ports than the target cell are rejected

memories with asynchronous read ports are rejected if not supported by target cell

a set of candidate “swizzle” transformations is assembled, for rearranging the memory geometry to match the target cell

the candidate swizzles are created according to target-specific rules, with some help from common functions for eg. legalizing write mask granularity

the best swizzle is selected, according to some metric (such as inserted soft mux depth or number of used target cells)

when a memory needs to be split/duplicated because of excessive read port count, each of the split/duplicated sub-memories will have its own swizzle

if a swizzle is egregiously bad (uses more resources than the entire FPGA has available), it is rejected

if applicable, a “fallback” lowering candidate is inserted into the list of candidates

no fallback lowering if multiple write port domains present

lowering to FFs for RAMs (very expensive)

lowering to LUTs for ROMs (somewhat expensive)

if no candidates are available, synthesis is aborted with a fatal error

one of the candidate lowerings is picked, according to some target-specific heuristic

initial heuristic will simply switch between target cells based on memory depth

in the future, we will want to take target resource counts into account (moving the depth threshold to avoid overfilling the FPGA)

if the “fallback” lowering candidate has been selected, the common lowering code is called to perform it

otherwise, target cell lowering is performed:

the memory is split into sub-memories (using common code), as required to satisfy max read port count, then each sub-memory is processed independently

various fixup transformations (implemented in common code, but driven by target code) are performed on the memory to match it with target cell capabilities:

sync read ports are converted to async if required

read port enable / clear / reset / initial value are unmapped and emulated if not supported

transparency is unmapped and emulated if not supported

read-first relations are emulated if not supported

the swizzle (computed earlier) is applied to the memory (using common code), aligning its geometry with the target cell

the memory is consumed and converted into actual target cells

any target-specific post-processing (such as merging output pipeline registers) is performed

Structs

MemorySwizzle

Describes a “swizzle” transformation, used to align memory geometry to target requirements.

Traits

MemoryExt

An extension trait for Memory with assorted memory lowering utility functions.