prjunnamed_memory
Trait MemoryExt

pub trait MemoryExt {
    // Required methods
    fn extract_read_ports(
        &self,
        ports: &[usize],
        output: &Value,
    ) -> (Memory, Value);
    fn unmap_read_dff(
        &mut self,
        design: &Design,
        port_index: usize,
        output: &mut Value,
    );
    fn unmap_read_init_reset_transparency(
        &mut self,
        design: &Design,
        port_index: usize,
        include_transparency: bool,
        output: &mut Value,
    );
    fn emulate_read_before_write(&mut self, design: &Design);
    fn make_data_swizzle(
        &self,
        write_port_granularity: &[usize],
    ) -> Vec<Option<usize>>;
    fn swizzle_depths(&self, swizzle: &MemorySwizzle) -> Vec<usize>;
    fn swizzle_mux_bits(
        &self,
        read_ports: &[usize],
        swizzle: &MemorySwizzle,
    ) -> (Vec<usize>, Vec<usize>);
    fn swizzle(
        self,
        design: &Design,
        output: &Value,
        swizzle: &MemorySwizzle,
    ) -> Vec<(Memory, Value)>;
    fn can_lower_fallback(&self) -> bool;
    fn lower_fallback(self, design: &Design, output: &Value);
}

An extension trait for Memory with assorted memory lowering utility functions.

Required Methods

fn extract_read_ports(&self, ports: &[usize], output: &Value) -> (Memory, Value)

Returns a new memory, with the same dimensions and write ports as this one, but only the selected subset of read ports.

output should be the output value of the memory. This function will slice out the bits corresponding to the selected read ports, and return them.

fn unmap_read_dff( &mut self, design: &Design, port_index: usize, output: &mut Value, )

Converts a synchronous read port to an asynchronous read port, by effectively extracting a flip-flop. If transparency relations are involved, soft transparency logic is inserted.

output is the output value of the memory cell. This function performs a replace_value on the bits corresponding to the converted port, and changes them in-place in the argument to newly-added void nets that are to be driven by the newly-converted port. Thus, the original memory cell must be unalived, and replace_value must be called on the new value of output, possibly after further lowering.

If the given read port is already asynchronous, the function will return without doing anything.

fn unmap_read_init_reset_transparency( &mut self, design: &Design, port_index: usize, include_transparency: bool, output: &mut Value, )

Unmaps the following features from a synchronous read port:

reset

clear

initial value

if include_transparency is true, all MemoryPortRelation::Transparent relations

If none of the above are used on the given port, the function will return without doing anything.

Note that, if any of the above features is unmapped, all of reset, clear, and initial value can be unmapped for free (it would actually be more complex to not unmap them). Thus, we do not have granular controls for them.

This function must not be called on asynchronous read ports.

fn emulate_read_before_write(&mut self, design: &Design)

Emulates all read-before-write relations in the memory.

This function has strict preconditions:

all read ports must be synchronous and have the same clock

a single write port cannot have both read-before-write and transparent relations

If there are no read-before-write relations in the memory, this function will return without doing anything.

The read-before-write relations are emulated by delaying all write ports with read-before-write relations by half a cycle: the clock polarity of the relevant write ports is flipped, and flip-flops (clocked by the original clock) are added on the data, mask, and address inputs.

fn make_data_swizzle( &self, write_port_granularity: &[usize], ) -> Vec<Option<usize>>

Constructs a data_swizzle for the MemorySwizzle structure, ensuring that the resulting memory will satisfy mask granularity constraints for all write ports. The argument must contain one number for each write port, specifying its mask granularity in bits. For each write port, the result will satisfy the following constraints:

the result’s length will be divisible by the specified granularity

every aligned group of granularity bits will consist only of bits that share the same mask net in the write port; if the write port is wide, this rule applies individually to all of its subports

fn swizzle_depths(&self, swizzle: &MemorySwizzle) -> Vec<usize>

Returns the depths of the memories that would be returned by swizzle, for cost estimation purposes. Each entry in the result vector corresponds to one memory.

For the simple case of the target mapping every memory to a single target cell, the length of the returned vector is equal to the number of target cells used. For more advanced mappings where the target can lower a memory to multiple target cells (e.g. where hardware cascade or chip selects are involved), the target can look at the exact memory depths required and perform its own computations.

fn swizzle_mux_bits( &self, read_ports: &[usize], swizzle: &MemorySwizzle, ) -> (Vec<usize>, Vec<usize>)

Returns the number of soft-decoded address bits for each write port and read port for a given swizzle.

Assumes that the swizzle applies to a memory where extract_read_ports has been called first with the given read_ports argument.

The first returned vector is the number of soft-decoded address bits for each write port, and the second is the number of soft-decoded address bits for each read port included in read_ports.

fn swizzle( self, design: &Design, output: &Value, swizzle: &MemorySwizzle, ) -> Vec<(Memory, Value)>

Performs a MemorySwizzle transformation on a memory, and returns a list of new memories with the corresponding outputs that need to be driven. The memories should be further lowered by the target.

fn can_lower_fallback(&self) -> bool

Returns true iff the memory is legal for fallback lowering.

A memory is legal for fallback lowering iff all write ports share the same clock.

fn lower_fallback(self, design: &Design, output: &Value)

Performs fallback lowering of the memory.

If the memory has write ports, a flip-flop is created for every bit of the memory, with associated write address decoding logic and read muxes. It is thus very inefficient unless the memory depth is very small.

If the memory has no write ports, it is lowered to just a tree of muxes for every read port, with constants at the leafs. On FPGA targets, this will be lowered to a fairly simple tree of LUTs, which can be reasonably efficient for memories of small depth.